Changes to Input handling
